Update lab1 and lab2 files

+++ lab-2/io.c
@@ -21,8 +21,8 @@ enum {
     Pull_Up = 1
 };
 
-void mmio_write(unsigned int reg, unsigned int val) { *(volatile unsigned int *)reg = val; }
-unsigned int mmio_read(unsigned int reg) { return *(volatile unsigned int *)reg; }
+void mmio_write(long reg, unsigned int val) { *(volatile unsigned int *)reg = val; }
+unsigned int mmio_read(long reg) { return *(volatile unsigned int *)reg; }
 
 unsigned int gpio_call(unsigned int pin_number, unsigned int value, unsigned int base, unsigned int field_size, unsigned int field_max) {
     unsigned int field_mask = (1 << field_size) - 1;
